requirements.txt
absl-py==2.1.0
accelerate==0.33.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.9.0
blis==0.7.11
cachetools==5.5.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.8
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
datasets==2.20.0
dill==0.3.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate==0.4.2
fastapi==0.115.11
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.1
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.5
idna==3.4
imageio==2.5.0
imantics==0.1.12
imbalanced-learn==0.12.2
imgaug==0.2.9
intel-openmp==2021.4.0
keras==3.5.0
keras-segmentation==0.3.0
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lime==0.2.0.1
marisa-trie==1.2.0
mkl==2021.4.0
ml-dtypes==0.3.2
mlxtend==0.23.1
mpmath==1.3.0
multiprocess==0.70.16
murmurhash==1.0.10
mysql-connector-python==8.1.0
namex==0.0.8
narwhals==1.23.0
nibabel==5.2.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
preshed==3.0.9
proto-plus==1.26.0
protobuf==4.21.12
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyngrok==7.2.3
pyparsing==3.2.1
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.4.2
segtok==1.5.11
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
skillsnetwork==0.21.9
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.46.1
streamlit==1.41.1
sympy==1.12
tbb==2021.13.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
thinc==8.2.5
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.5
transformers==4.44.1
typer==0.12.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.5
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
xgboost==2.0.3
xmljson==0.2.1
xxhash==3.4.1
yake==0.4.8
